// src/store/types/TTWSlice.tsx
export type TTWAddFormResponse = {
  Selected_Idv_Default: string;
  Plans_Listed_Final: string;
  lead_id: string;
  num_brands_listed: number;
  num_plans_listed: number;
  quote_no: string;
  module: string;
  insurance_type: string;
  name: string;
  email: string;
  mobile: string;
  city: string;
  state: string;
  utm_source: string;
  utm_medium: string;
  utm_medium_m: string;
  utm_medium_d: string;
  utm_campaign: string;
  utm_keyword: string;
  status: number;
  device: string;
  ip_address: string;
  lead_for: number;
  make_model: string;
  variant_cc: string;
  fuel_type: string;
  business_type: TBusinessType;
  tp_policy_start_date: string;
  policy_expiry_date: string;
  new_policy_date: string;
  manufacture_date: string;
  registration_date: string;
  registration_no: string;
  actual_reg_no: string;
  rto_code: string;
  rto_location: string;
  previous_insurer: string;
  tp_previous_insurer: string;
  previous_ncb: TNCB;
  new_ncb: TNCB;
  claim_made: "Yes" | "No";
  electric_accessory: number;
  non_electric_accessory: number;
  passernger_cover: number;
  cng_kit: number;
  deductible: number;
  driving_exp: number;
  avg_monthly_milege: number;
  anti_theft_device: string;
  automobile_association: string;
  ambulance_charges_cover: number;
  consumables_cover: number;
  depreciation_cover: number;
  hydrostatic_lock_cover: number;
  medical_expenses: number;
  invoice_cover: number;
  road_side_assistance: number;
  engine_gear_box_protection: number;
  no_claim_bonus_same_slab: number;
  driver_cover: number;
  pa_owner_driver: number;
  reason_for_exemption_pa: number;
  vehicle_owned: string;
  owner_occupation: string;
  owner_dob: string;
  bundled_policy: number;
  idv: number;
  tw_master_data_id: number;
  make_id: number;
  model_id: number;
  variant_id: number;
  api_unique_id: string;
  make: string;
  model: string;
  variant: string;
  cubic_capacity: string;
  seating_capacity: string;
  api_owner_name: string;
  api_owner_gender: string;
  api_owner_email: string;
  api_owner_dob: string;
  api_registration_date: string;
  api_chasis_no: string;
  api_engine_no: string;
  api_vehicle_class: string;
  api_maker: string;
  api_rto: string;
  api_address_p_house: string;
  api_address_p_pincode: string;
  api_address_p_area: string;
  api_address_p_district: string;
  api_address_p_city: string;
  api_address_p_state: string;
  requestTime: number;
  header_code_id: number;
  header_code_desc: string;
  insurer_types: {};
};

export type TTWForm = {
  reg_no: TFieldData;
  invoice_date: TFieldData;
  reg_date: TFieldData;
  business_type: "Rollover" | "New" | "Expired";
  make: TFieldData;
  model: TFieldData;
  make_model: TFieldData;
  fuel_type: TFieldData;
  variant_cc: TFieldData;
  previous_insurer: TFieldData;
  mobile: TFieldData;
  claimed: "Yes" | "No";
  prev_ncb: TNCB;
  rto: TFieldData;
  policy_expiry_date: TFieldData;
  name: TFieldData;
  previous_policy_type: TFieldData;
  utm_medium_d: string;
  pincode: TFieldData;
};

export type TWDropdown = {
  OWNER_MARTIAL_STATUS_DATA: TDropdown[];
  CLAIM_STATUS: TDropdown[];
  POLICY_TYPE: TDropdown[];
  POPULAR_MAKE_LIST: TDropdown[];
  PLAN_TYPE: TDropdown[];
  SORT_BY: TDropdown[];
  MAKE_MODEL_LIST: TDropdown[];
  RTO_LIST: TDropdown[];
  MAKE_LIST: TDropdown[];
  MODEL_LIST: TDropdown[];
  FUEL_TYPE_LIST: TDropdown[];
  VARIANT_LIST: TDropdown[];
  PREVIOUS_INSURER_LIST: TDropdown[];
  NCB: TDropdown[];
  POPULAR_MAKE: string[];
  VEHICLE_ON_LOAN: string[];
  PA_OWNER_COVER_DATA: TDropdown[];
  NOMINEE_RELATIONSHIP: TDropdown[];
};

export type TVehicleDetails = {
  reg_date: TFieldData;
  chasis_no: TFieldData;
  engine_no: TFieldData;
  prev_policy_no: TFieldData;
  prev_insurer: TFieldData;
  mfg_year: TFieldData;
  policy_exp_date: TFieldData;
  tp_policy: TFieldData;
  tp_insurer: TFieldData;
  tp_policy_start: TFieldData;
  bank_name: TFieldData;
  city: TFieldData;
  is_vehicle_on_loan: boolean;
  invoice_date: TFieldData;
  PAN: TFieldData;
  IS_OWNERSHIP_CHANGED: boolean;
  prev_depth_cover: boolean;
};
export type TCkycDetails = {
  propFullName: string;
  ProposerDob: string;
  docID: string;
  cusdtomertype: string;
  verificationDoc: string;
  gender: string;
};

export type TOwnerDetails = {
  pincode?: { value: string; warning: boolean };
  name: { value: string; warning: boolean };
  mobile: { value: string; warning: boolean };
  email: { value: string; warning: boolean };
  dob: TFieldData;
  gender: { value: string; warning: boolean };
  maritalStatus: { value: string; warning: boolean };
  pacover: { value: string; warning: boolean };
  GST: { value: string; warning: boolean };
  validPuc: boolean;
  vehicleRegistered: boolean;
  isPolicyComprehensive: boolean;
  includePAOwner: boolean;
  customer_type: TFieldData;
};

export type DROPDOWN_DATA = {
  RTO_LIST: TDropdown[];
  MAKE_LIST: TDropdown[];
  MODEL_LIST: TDropdown[];
  MAKE_MODEL_LIST: TDropdown[];
  FUEL_TYPE_LIST: TDropdown[];
  VARIANT_LIST: TDropdown[];
  PREVIOUS_INSURER_LIST: TDropdown[];
  NCB: string[];
  POPULAR_MAKE: string[];
  PASSENGER_COVER_DATA: TDropdown[];
  PLAN_TYPE: TDropdown[];
  SORT_BY: TDropdown[];
};
export type TNomineeDetails = {
  fullName: TFieldData;
  dob: TFieldData;
  relationship: TFieldData;
};
export type IDV_VALUE = {
  value: string;
  warning: boolean;
};
export type TTWPlanType =
  | "Third Party"
  | "Comprehensive"
  | "Bundle Policy"
  | "SAOD"
  | "";
export type TTWSlice = {
  REDIRECTED: boolean;
  ADD_FORM: TTWForm;
  ADD_FORM_RESPONSE: TTWAddFormResponse;
  DROPDOWN_DATA: TWDropdown;
  ADDON_STATUS: TTWAddonStatus;
  IDV_VALUE: IDV_VALUE;
  PRODUCT_LIST: TTWProduct[];
  QUOTE_LOADER: boolean;
  ALL_QUOTES_DATA: TTWProduct[];
  SELECTED_QUOTE_DATA: TTWProduct;
  create_proposal_response: any;
  policyId: any;
  PLAN_TYPE: TTWPlanType;
  SORT_BY: "Low to high" | "High to low";
  IDV_SELECTED_OPTION: "Best Deal" | "Lowest IDV" | string;
  OWNER_DETAILS: TOwnerDetails;
  VEHICLE_DETAILS: TVehicleDetails;
  CKYC_DETAILS: TCkycDetails;
  NOMINEE_DETAILS: TNomineeDetails;
  ADDRESS_DETAILS: TAddressDetails;
  IDVApply_Status: boolean;
  PAGE_STATUS: boolean;
  Home_Loader: boolean;
  CUS_TYPE: boolean;
  PLAN_DETAILS: TTWPlanDetails;
  SHOW_EDIT_PLAN_DETAILS_POPUP: boolean;
  SHOW_PRICE_DETAILS_POPUP: boolean;
  NO_QUOTE_REASONS: TTWNoQuoteReason[];
};